refactor: extract entity-select components into module
for easier overview of related components and simplified testing imports
