+++ src/app/child-dev-project/children/children.module.ts
@@ -66,49 +66,51 @@ import { ExportModule } from "../../core/export/export.module";
 import { BirthdayDashboardComponent } from "./dashboard-widgets/birthday-dashboard/birthday-dashboard.component";
 import { ChildBlockTooltipComponent } from "./child-block/child-block-tooltip/child-block-tooltip.component";
 import { CommonComponentsModule } from "../../core/common-components/common-components.module";
+import { EntitySelectModule } from "../../core/entity-components/entity-select/entity-select.module";
 
 @NgModule({
-  imports: [
-    CommonModule,
-    FormsModule,
-    RouterModule,
-    MatFormFieldModule,
-    MatInputModule,
-    MatCheckboxModule,
-    MatSelectModule,
-    MatDatepickerModule,
-    MatExpansionModule,
-    MatTableModule,
-    MatPaginatorModule,
-    MatSortModule,
-    MatSidenavModule,
-    MatButtonModule,
-    MatButtonToggleModule,
-    MatCardModule,
-    MatSnackBarModule,
-    MatProgressBarModule,
-    MatAutocompleteModule,
-    MatTooltipModule,
-    MatChipsModule,
-    BrowserAnimationsModule,
-    ReactiveFormsModule,
-    SchoolsModule,
-    ReactiveFormsModule,
-    MatDialogModule,
-    MatListModule,
-    MatProgressSpinnerModule,
-    ConfirmationDialogModule,
-    FormDialogModule,
-    Angulartics2Module,
-    ViewModule,
-    EntitySubrecordModule,
-    EntityListModule,
-    EntityUtilsModule,
-    DashboardModule,
-    FontAwesomeModule,
-    ExportModule,
-    CommonComponentsModule,
-  ],
+    imports: [
+        CommonModule,
+        FormsModule,
+        RouterModule,
+        MatFormFieldModule,
+        MatInputModule,
+        MatCheckboxModule,
+        MatSelectModule,
+        MatDatepickerModule,
+        MatExpansionModule,
+        MatTableModule,
+        MatPaginatorModule,
+        MatSortModule,
+        MatSidenavModule,
+        MatButtonModule,
+        MatButtonToggleModule,
+        MatCardModule,
+        MatSnackBarModule,
+        MatProgressBarModule,
+        MatAutocompleteModule,
+        MatTooltipModule,
+        MatChipsModule,
+        BrowserAnimationsModule,
+        ReactiveFormsModule,
+        SchoolsModule,
+        ReactiveFormsModule,
+        MatDialogModule,
+        MatListModule,
+        MatProgressSpinnerModule,
+        ConfirmationDialogModule,
+        FormDialogModule,
+        Angulartics2Module,
+        ViewModule,
+        EntitySubrecordModule,
+        EntityListModule,
+        EntityUtilsModule,
+        DashboardModule,
+        FontAwesomeModule,
+        ExportModule,
+        CommonComponentsModule,
+        EntitySelectModule,
+    ],
   declarations: [
     ChildBlockComponent,
     ChildrenListComponent,
