+++ src/app/child-dev-project/notes/notes.module.ts
@@ -48,6 +48,7 @@ import { DashboardModule } from "../../core/dashboard/dashboard.module";
 import { ExportModule } from "../../core/export/export.module";
 import { CommonComponentsModule } from "../../core/common-components/common-components.module";
 import { ImportantNotesDashboardComponent } from "./dashboard-widgets/important-notes-dashboard/important-notes-dashboard.component";
+import { EntitySelectModule } from "../../core/entity-components/entity-select/entity-select.module";
 
 @NgModule({
   declarations: [
@@ -59,55 +60,56 @@ import { ImportantNotesDashboardComponent } from "./dashboard-widgets/important-
     NotesRelatedToEntityComponent,
     ImportantNotesDashboardComponent,
   ],
-  imports: [
-    CommonModule,
-    FormsModule,
-    ConfirmationDialogModule,
-    RouterModule,
-    MatFormFieldModule,
-    MatInputModule,
-    MatCheckboxModule,
-    MatSelectModule,
-    MatDatepickerModule,
-    MatExpansionModule,
-    MatTableModule,
-    MatSortModule,
-    MatSidenavModule,
-    MatButtonModule,
-    MatButtonToggleModule,
-    MatCardModule,
-    MatSnackBarModule,
-    MatDialogModule,
-    MatProgressBarModule,
-    MatAutocompleteModule,
-    MatTooltipModule,
-    MatTabsModule,
-    MatSlideToggleModule,
-    BrowserAnimationsModule,
-    ReactiveFormsModule,
-    SchoolsModule,
-    ReactiveFormsModule,
-    MatDialogModule,
-    MatListModule,
-    ChildrenModule,
-    DragDropModule,
-    MatTreeModule,
-    MatTabsModule,
-    MatPaginatorModule,
-    FormDialogModule,
-    Angulartics2Module,
-    EntitySubrecordModule,
-    EntityListModule,
-    ConfigurableEnumModule,
-    AttendanceModule,
-    MatSlideToggleModule,
-    EntityUtilsModule,
-    FontAwesomeModule,
-    MatMenuModule,
-    DashboardModule,
-    ExportModule,
-    CommonComponentsModule,
-  ],
+    imports: [
+        CommonModule,
+        FormsModule,
+        ConfirmationDialogModule,
+        RouterModule,
+        MatFormFieldModule,
+        MatInputModule,
+        MatCheckboxModule,
+        MatSelectModule,
+        MatDatepickerModule,
+        MatExpansionModule,
+        MatTableModule,
+        MatSortModule,
+        MatSidenavModule,
+        MatButtonModule,
+        MatButtonToggleModule,
+        MatCardModule,
+        MatSnackBarModule,
+        MatDialogModule,
+        MatProgressBarModule,
+        MatAutocompleteModule,
+        MatTooltipModule,
+        MatTabsModule,
+        MatSlideToggleModule,
+        BrowserAnimationsModule,
+        ReactiveFormsModule,
+        SchoolsModule,
+        ReactiveFormsModule,
+        MatDialogModule,
+        MatListModule,
+        ChildrenModule,
+        DragDropModule,
+        MatTreeModule,
+        MatTabsModule,
+        MatPaginatorModule,
+        FormDialogModule,
+        Angulartics2Module,
+        EntitySubrecordModule,
+        EntityListModule,
+        ConfigurableEnumModule,
+        AttendanceModule,
+        MatSlideToggleModule,
+        EntityUtilsModule,
+        FontAwesomeModule,
+        MatMenuModule,
+        DashboardModule,
+        ExportModule,
+        CommonComponentsModule,
+        EntitySelectModule,
+    ],
   exports: [NoteDetailsComponent],
 })
 export class NotesModule {
