+++ src/app/core/entity-components/entity-details/form/form.component.spec.ts
@@ -5,6 +5,7 @@ import { Child } from "../../../../child-dev-project/children/model/child";
 import { Router } from "@angular/router";
 import { MockedTestingModule } from "../../../../utils/mocked-testing.module";
 import { EntityDetailsModule } from "../entity-details.module";
+import { ConfirmationDialogService } from "../../../confirmation-dialog/confirmation-dialog.service";
 
 describe("FormComponent", () => {
   let component: FormComponent;
@@ -13,6 +14,7 @@ describe("FormComponent", () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
       imports: [EntityDetailsModule, MockedTestingModule.withState()],
+      providers: [{ provide: ConfirmationDialogService, useValue: null }],
     }).compileComponents();
   });
