+++ src/app/core/entity-components/entity-select/display-entity-array/display-entity-array.component.html
