+++ src/app/core/entity-components/entity-select/display-entity-array/display-entity-array.stories.ts
@@ -1,15 +1,15 @@
 import { Story, Meta } from "@storybook/angular/types-6-0";
 import { moduleMetadata } from "@storybook/angular";
-import { Child } from "../../../../../child-dev-project/children/model/child";
-import { SchoolsModule } from "../../../../../child-dev-project/schools/schools.module";
-import { ChildrenModule } from "../../../../../child-dev-project/children/children.module";
+import { Child } from "../../../../child-dev-project/children/model/child";
+import { SchoolsModule } from "../../../../child-dev-project/schools/schools.module";
+import { ChildrenModule } from "../../../../child-dev-project/children/children.module";
 import { DisplayEntityArrayComponent } from "./display-entity-array.component";
 import { BehaviorSubject } from "rxjs";
-import { StorybookBaseModule } from "../../../../../utils/storybook-base.module";
-import { EntityUtilsModule } from "../../entity-utils.module";
-import { EntityMapperService } from "../../../../entity/entity-mapper.service";
-import { mockEntityMapper } from "../../../../entity/mock-entity-mapper-service";
-import { ChildrenService } from "../../../../../child-dev-project/children/children.service";
+import { StorybookBaseModule } from "../../../../utils/storybook-base.module";
+import { EntityUtilsModule } from "../../entity-utils/entity-utils.module";
+import { EntityMapperService } from "../../../entity/entity-mapper.service";
+import { mockEntityMapper } from "../../../entity/mock-entity-mapper-service";
+import { ChildrenService } from "../../../../child-dev-project/children/children.service";
 
 const child1 = new Child();
 child1.name = "Test Name";
