+++ src/app/core/entity-components/entity-select/display-entity/display-entity.component.html
