+++ src/app/core/entity-components/entity-select/edit-entity-array/edit-entity-array.component.spec.ts
@@ -1,10 +1,10 @@
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 
 import { EditEntityArrayComponent } from "./edit-entity-array.component";
-import { Child } from "../../../../../child-dev-project/children/model/child";
-import { EntityUtilsModule } from "../../entity-utils.module";
-import { setupEditComponent } from "../edit-component.spec";
-import { MockedTestingModule } from "../../../../../utils/mocked-testing.module";
+import { Child } from "../../../../child-dev-project/children/model/child";
+import { setupEditComponent } from "../../entity-utils/dynamic-form-components/edit-component.spec";
+import { MockedTestingModule } from "../../../../utils/mocked-testing.module";
+import { EntitySelectModule } from "../entity-select.module";
 
 describe("EditEntityArrayComponent", () => {
   let component: EditEntityArrayComponent;
@@ -12,7 +12,7 @@ describe("EditEntityArrayComponent", () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [EntityUtilsModule, MockedTestingModule.withState()],
+      imports: [EntitySelectModule, MockedTestingModule.withState()],
     }).compileComponents();
   });
