+++ src/app/core/entity-components/entity-select/edit-entity-array/entity-reference-array.stories.ts
@@ -1,18 +1,18 @@
 import { Story, Meta } from "@storybook/angular/types-6-0";
 import { moduleMetadata } from "@storybook/angular";
-import { EntityUtilsModule } from "../../entity-utils.module";
-import { EntitySchemaService } from "../../../../entity/schema/entity-schema.service";
-import { EntityFormComponent } from "../../../entity-form/entity-form/entity-form.component";
-import { EntityFormModule } from "../../../entity-form/entity-form.module";
-import { FormFieldConfig } from "../../../entity-form/entity-form/FormConfig";
-import { EntityMapperService } from "../../../../entity/entity-mapper.service";
-import { AlertsModule } from "../../../../alerts/alerts.module";
-import { StorybookBaseModule } from "../../../../../utils/storybook-base.module";
-import { DatabaseEntity } from "../../../../entity/database-entity.decorator";
-import { Entity } from "../../../../entity/model/entity";
-import { DatabaseField } from "../../../../entity/database-field.decorator";
-import { mockEntityMapper } from "../../../../entity/mock-entity-mapper-service";
-import { User } from "../../../../user/user";
+import { EntityUtilsModule } from "../../entity-utils/entity-utils.module";
+import { EntitySchemaService } from "../../../entity/schema/entity-schema.service";
+import { EntityFormComponent } from "../../entity-form/entity-form/entity-form.component";
+import { EntityFormModule } from "../../entity-form/entity-form.module";
+import { FormFieldConfig } from "../../entity-form/entity-form/FormConfig";
+import { EntityMapperService } from "../../../entity/entity-mapper.service";
+import { AlertsModule } from "../../../alerts/alerts.module";
+import { StorybookBaseModule } from "../../../../utils/storybook-base.module";
+import { DatabaseEntity } from "../../../entity/database-entity.decorator";
+import { Entity } from "../../../entity/model/entity";
+import { DatabaseField } from "../../../entity/database-field.decorator";
+import { mockEntityMapper } from "../../../entity/mock-entity-mapper-service";
+import { User } from "../../../user/user";
 
 const testUser = new User("1");
 testUser.name = "test entity";
