+++ src/app/core/entity-components/entity-select/edit-single-entity/edit-single-entity.component.html
