+++ src/app/core/entity-components/entity-select/edit-single-entity/edit-single-entity.component.scss
@@ -1,4 +1,4 @@
-@use "../../common-styles" as *;
+@use "../../entity-utils/common-styles" as *;
 
 .block-wrapper {
   @include entity-block-border;
