+++ src/app/core/entity-components/entity-select/edit-text-with-autocomplete/edit-text-with-autocomplete.component.scss
