+++ src/app/core/entity-components/entity-select/entity-select/entity-select.component.scss
@@ -1,4 +1,4 @@
-@use "../common-styles" as *;
+@use "../../entity-utils/common-styles" as *;
 
 .main-list {
   // Mat chip list overflows a little bit without this hack
