+++ src/app/core/ui/ui.module.ts
@@ -42,36 +42,38 @@ import { LanguageModule } from "../language/language.module";
 import { FontAwesomeModule } from "@fortawesome/angular-fontawesome";
 import { PwaInstallModule } from "app/core/pwa-install/pwa-install.module";
 import { UserSecurityComponent } from "../user/user-security/user-security.component";
+import { EntitySelectModule } from "../entity-components/entity-select/entity-select.module";
 
 /**
  * The core user interface structure that ties different components together into the overall app layout.
  */
 @NgModule({
-  imports: [
-    CommonModule,
-    FormsModule,
-    LatestChangesModule,
-    NavigationModule,
-    RouterModule,
-    SessionModule,
-    SyncStatusModule,
-    MatSidenavModule,
-    MatToolbarModule,
-    MatButtonModule,
-    MatFormFieldModule,
-    MatInputModule,
-    MatAutocompleteModule,
-    ChildrenModule,
-    SchoolsModule,
-    NotesModule,
-    Angulartics2Module,
-    PermissionsModule,
-    EntityUtilsModule,
-    ReactiveFormsModule,
-    LanguageModule,
-    FontAwesomeModule,
-    PwaInstallModule,
-  ],
+    imports: [
+        CommonModule,
+        FormsModule,
+        LatestChangesModule,
+        NavigationModule,
+        RouterModule,
+        SessionModule,
+        SyncStatusModule,
+        MatSidenavModule,
+        MatToolbarModule,
+        MatButtonModule,
+        MatFormFieldModule,
+        MatInputModule,
+        MatAutocompleteModule,
+        ChildrenModule,
+        SchoolsModule,
+        NotesModule,
+        Angulartics2Module,
+        PermissionsModule,
+        EntityUtilsModule,
+        ReactiveFormsModule,
+        LanguageModule,
+        FontAwesomeModule,
+        PwaInstallModule,
+        EntitySelectModule,
+    ],
   declarations: [SearchComponent, UiComponent, PrimaryActionComponent],
   exports: [UiComponent],
 })
